+++ lib86cpu/core/translate.cpp
@@ -1023,7 +1023,7 @@ tc_unlink_all(cpu_t *cpu)
 }
 
 template<bool remove_hook>
-void tc_invalidate(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint8_t size)
+void tc_invalidate(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint64_t size)
 {
 	bool halt_tc = false;
 
@@ -1100,8 +1100,8 @@ void tc_invalidate(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint8_
 	}
 }
 
-template void tc_invalidate<true>(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint8_t size);
-template void tc_invalidate<false>(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint8_t size);
+template void tc_invalidate<true>(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint64_t size);
+template void tc_invalidate<false>(cpu_ctx_t *cpu_ctx, addr_t phys_addr, [[maybe_unused]] uint64_t size);
 
 static translated_code_t *
 tc_cache_search(cpu_t *cpu, addr_t pc)
